feat: [S12P21A508-160] long을 int로 변경

+++ backend/statics-server/src/main/java/com/ssafy/staticsserver/interfaces/country/dto/NewsModalResponse.java
@@ -11,7 +11,7 @@ public class NewsModalResponse {
 	private List<NewsDto> news;
 	private int page;
 	private int size;
-	private long totalElements;
+	private int totalElements;
 	private int totalPages;
 	private boolean hasNext;
 	private boolean hasPrevious;
